\name{SWSamp-package}
\alias{SWSamp-package}
\alias{SWSamp}
\docType{package}
\title{
SWSamp
}
\description{
Sample size calculations for a Stepped Wedge Trial
}
\details{
\tabular{ll}{
Package: \tab SWSamp\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2016-08-17\cr
License: \tab GPL2 \cr
LazyLoad: \tab yes\cr
}
The package provides a suite of function to compute the power for a Stepped Wedge Design
under different assumptions. The package can generate power based on simulations or use
closed-formulae based on Hussey et Hughes
}
\author{
Gianluca Baio

Maintainer: Gianluca Baio (g.baio@ucl.ac.uk)
}
\references{
Baio, G; Copas, A; Ambler, G; Hargreaves, J; Beard, E; and Omar, RZ Sample size calculation for 
a stepped wedge trial. Trials, 16:354. Aug 2015.

Hussey M and Hughes J. Design and analysis of stepped wedge cluster randomized trials.
Contemporary Clinical Trials. 28(2):182-91. Epub 2006 Jul 7. Feb 2007
}
\keyword{Sample size calculations}
\keyword{Stepped Wedge Design}

