% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSamp-package.R
\docType{package}
\name{SWSamp-package}
\alias{SWSamp}
\alias{SWSamp-package}
\title{SWSamp: Computes Sample Size for a Stepped Wedge Design, using Simulation-Based Calculations}
\description{
Provides a suite of functions for the sample size calculations and power analysis in a Stepped Wedge Trial. Contains functions for closed-form sample size calculation (based on a set of specific models) and simulation-based procedures that can extend the basic framework.
}
\author{
\strong{Maintainer}: Gianluca Baio \email{g.baio@ucl.ac.uk} (\href{https://orcid.org/0000-0003-4314-2570}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Rosie Leech
}

}
\keyword{internal}
